requirements.txt
Flask>=3.0.0
Flask-CORS>=4.0.0
SQLAlchemy>=2.0.35
openai>=1.12.0
anthropic>=0.18.1
google-genai>=1.0.0
python-dotenv>=1.0.0
